requirements.txt
opencv-python
transformers==4.32.0
modelscope
flask
gunicorn
accelerate
boto3
tiktoken
einops
transformers_stream_generator==0.0.4
scipy
torchvision
pillow
tensorboard
matplotlib
Flask==3.1.1
torch==2.7.0+cu118
transformers==4.53.0
opencv-python==4.11.0.86
Pillow==11.2.1
numpy==2.2.6
regex==2024.11.6
requests==2.32.3
Werkzeug==3.1.3